# server/model/community/response/community_captcha.go
package response

type CaptchaResponse struct {
	CaptchaId     string `json:"captchaId"`
	PicPath       string `json:"picPath"`
	CaptchaLength int    `json:"captchaLength"`
	OpenCaptcha   bool   `json:"openCaptcha"`
}

type CodeResponse struct {
	Phone string `json:"phone"`
	Code  string `json:"code"`
}